+++ src/components/Message/__tests__/MessageTeam.test.js
@@ -103,12 +103,6 @@ describe('<MessageTeam />', () => {
   afterEach(cleanup);
   beforeEach(jest.clearAllMocks);
 
-  it('should not render anything if message is of type message.read', async () => {
-    const message = generateAliceMessage({ type: 'message.read' });
-    const { container } = await renderMessageTeam(message);
-    expect(container).toBeEmptyDOMElement();
-  });
-
   it('should render deleted message with custom component when message was deleted and a custom delete message component was passed', async () => {
     const deletedMessage = generateAliceMessage({
       deleted_at: new Date('2019-08-27T00:24:00'),